% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/operators.R
\name{insertCompoundAssignmentOperator}
\alias{insertCompoundAssignmentOperator}
\title{Insert magrittr's compound assignment operator \%<>\%}
\usage{
insertCompoundAssignmentOperator()
}
\description{
Call this function as an addin to insert \code{\link[magrittr]{\%<>\%}} at the cursor position.
}
