% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/browser.R
\name{insertDebugBrowser}
\alias{insertDebugBrowser}
\title{Insert a call to browser()}
\usage{
insertDebugBrowser()
}
\description{
Call this function as an addin to insert `browser()` at the cursor position.
}
