% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/operators.R
\name{insertExpositionOperator}
\alias{insertExpositionOperator}
\title{Insert magrittr's exposition operator \%$\%}
\usage{
insertExpositionOperator()
}
\description{
Call this function as an addin to insert \code{\link[magrittr]{\%$\%}} at the cursor position.
}
