% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/operators.R
\name{insertRightAssignmentOperator}
\alias{insertRightAssignmentOperator}
\title{Insert left-to-right assignment operator ->}
\usage{
insertRightAssignmentOperator()
}
\description{
Call this function as an addin to insert \code{->} at the cursor position.
}
