% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/operators.R
\name{insertTeeOperator}
\alias{insertTeeOperator}
\title{Insert magrittr's tee operator \%T>\%}
\usage{
insertTeeOperator()
}
\description{
Call this function as an addin to insert \code{\link[magrittr]{\%T>\%}} at the cursor position.
}
