% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/readKnitrCache.R
\name{readKnitrCache}
\alias{readKnitrCache}
\title{Read cached chunks from this .Rmd}
\usage{
readKnitrCache()
}
\description{
Read cached chunks from this .Rmd. All variables are loaded into the global environment.
}
