% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/viewer.R
\name{rs_get_symbol_at_cursor}
\alias{rs_get_symbol_at_cursor}
\title{Get the symbol under cursor (as character vector).}
\usage{
rs_get_symbol_at_cursor(context)
}
\value{
a character vector
}
\description{
Get the symbol under cursor (as character vector).
From `ropensci/drake`.
}
