% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/runShinyMarkdown.R
\name{runShinyMarkdown}
\alias{runShinyMarkdown}
\title{Run Shiny document in current session}
\usage{
runShinyMarkdown()
}
\description{
Run Shiny document in current session instead of render-ing to a separate session.
This enables browser() calls.
}
