% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/sourceKnitr.R
\name{sourceKnitr}
\alias{sourceKnitr}
\title{Source the current .Rmd file}
\usage{
sourceKnitr()
}
\description{
Source the current .Rmd file as if it were a R script.
}
\details{
All variables are loaded into the global environment.
}
